Create SFP types file
